cleanup imports and move config
